refactor: url_query_helper
